A very rough sketch of what would be a PER-based construction of ℝ. No details, very ugly, sorry and all that.
There are still a lot of axioms to prove (see Raxioms.v) and 
a lot of compatibility to state (as you'll see the 
compatibility of proof-depedent functions like "Qinv" is 
very hard to state, I didn't expect that before writing it, 
couldn't we find a few syntactic sugar in order to facilitate
this ?). 

I fear that this little piece of code will convince you that
setoids are not good for us... even if my goal was to show
how convenient they are :). 

(By the way, guys, you can finish this sketch for me, it's 
actually quite fun. The last two proofs are a good exercise :)). 

`Make Rcauchy.vo` to compile it.



git-svn-id: svn+ssh://svn.code.sf.net/p/coqtail/code@149 7194d10d-a3ed-4628-aea2-7b4a0dfde000