======================== Big update: Rpser ========================
==============================
========= New things =========
==============================

All the modifications are done in order to have shorter proofs by
reflecting the structure behind the objects.

-- gt_pser, gt_abs_pser, An_deriv, etc. are defined in terms of
operations on Rsequences. Rpser_def_simpl gives rewriting rules
on these new abstractions;

-- Partial power series are special objects (Rseq_pps) defined using
Rseq_sum and gt_pser;

-- We do not rely on Un_cv, infinite_sum, Pser, etc. anymore except
for the RFsequence part which should be modified;

==============================
========== Updates ===========
==============================

-- everything in Rpser is up-to-date;

-- some parts of Cpser have been made compilable but the code is
really ugly and does not use the new ideas;

==============================
====== Broken libraries ======
==============================

I have updated a lot of things this week-end but there was too
much to do. Most of the time adding a couple of [unfold],
modifying the [Require Import] list to add the new libraries
and changing the names that have changed  is sufficient.
Sometimes it is more subtle (e.g. on the complex where there are
collisions between IRC / Cmult / Rmult).

-- RTaylor, RStirling, Rzeta2, Triangular, Wallis are broken

-- Cexp is broken;



========================= Everything else =========================

-- README records the revision number before the update (where
everything works fine)

-- Differential equations can now contain a multiplication by a
constant

-- mytheories contains new stuff (nothing fancy)

-- Rsequence_rewrite_facts is basically lemmas that should maybe
be moved somewhere appropriate



git-svn-id: svn+ssh://svn.code.sf.net/p/coqtail/code@269 7194d10d-a3ed-4628-aea2-7b4a0dfde000