On the way to a quote function.
Following one of the remarks from the anonymous
referees: you cannot really talk about reflection
unless you are able to reify your goal.

Optimizations can be discussed:

- the fact that we do not care about the shape of the
[infinite_cv_radius an] proof
- the fact that we do not care about the shape of the
[C n (sum an rho)] proof (shouldn't we use [D n] instead
of [C n]?! Scandal!)
- the fact that [isconst s x] is a depth-first from right
to left algorithm

By the way: nothing new but debugging Ltac in a shell is
really a pain in the ass.


git-svn-id: svn+ssh://svn.code.sf.net/p/coqtail/code@286 7194d10d-a3ed-4628-aea2-7b4a0dfde000