Adding An_expand to solve_diff_equa. Cool stuff happening: one can quote / solve [Rexp (a*x) * Rexp (b*x) = Rexp ((a+b)*b)]
This proof is however maybe not the most interesting one (instead
of brute-forcing on the coefficients of the power series, one could
prove that both sides have the same derivative and that they have
the same value on 0).

BUT it can be easily (ouececa) reduced to the Newton's binomial
theorem. Is it proved somewhere? Jm?


git-svn-id: svn+ssh://svn.code.sf.net/p/coqtail/code@308 7194d10d-a3ed-4628-aea2-7b4a0dfde000