Introducing: classes on restriction to balls. This is however really experimental and will probably change in the near future.
A couple of remarks:

- the continuity is really cleanly defined (continuity_pt,
continue_in, limit1_in, limit_in) whereas the derivability
is somehow ad-hoc and non modular

- continuity_Rball will probably stay unchanged (apart from
D_x that should maybe be removed): when we talk about the
continuity inside a ball, we consider only points inside this
ball (thanks to continue_in).

- modifying interval_continuity in the same fashion would break
RIVT and I do not want to have to fix it right now. But it would
make sense to do the same (or develop a similar and somehow cleaner
library for continuity_Rball and then use conversions between
continuity_Rball and interval_continuity to retrieve the results).



git-svn-id: svn+ssh://svn.code.sf.net/p/coqtail/code@315 7194d10d-a3ed-4628-aea2-7b4a0dfde000