Bug fix: clusters' names are now unique integers.
With the previous implementation, all the files in the
subdirectory [XXX] where gathered in a cluster named
[clusterXXX].

In coqtail we have the directory [src/Reals/] as well as
[src/Fresh/Reals] and we ended up producing a dot file
with two distinct clusters named [clusterReals].
And this was confusing for dot.

Now that each name is unique, we won't have this kind of
problem anymore.

** Special thanks **
Bug fixed thanks to marapet's diagnostic on stackoverflow:
http://stackoverflow.com/questions/11584771/cryptic-dot-error-message-for-graph-with-big-subclusters/



git-svn-id: svn+ssh://svn.code.sf.net/p/coqtail/code@401 7194d10d-a3ed-4628-aea2-7b4a0dfde000